Migrate to sidebar layout
